Add possibility to dynamically update the sunrise and sunset times in a scheduled task.
This PR targets the WIP branch as it is definitely
considered to still be a WIP after it was merged.

It makes sheduled tasks finally work. And it adds the possibility to
register a task that dynamically updates the sunrise and sunset times
based on the currently active time zone and the specified geographic
coordinates. For the latter, algorithms from Paul Schlyter's SUNRISET.C
are taken. https://stjarnhimlen.se/comp/sunriset.c