Instantly run if time triggers are overdue. (#5)
The docs state:
> Tasks that are started after the scheduled time has passed (because
> of the StartWhenAvailable property being set to True) are queued in
> the Task Scheduler service's queue of tasks and they are started
> after a delay. The default delay is 10 minutes.
 
That means, if the computer is started after sunrise and the mode was
dark when it has been shut down, it takes 10 minutes before the mode
is turned to light. That's unexpected for the purpose of this app.
This behavior doesn't change if the priority of the task is updated
from below normal (default) to normal. It turns out that the
`MissedTasksStartupDelay` value in registry key
```
HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Schedule\Configuration
```
defines the delay. Updating this value is no option because
- all tasks are affected
- updates in `HKLM` require elevation

The easiest way to get around this is adding a logon trigger and a
resume-from-hibernation event trigger. The `StartWhenAvailable`
setting isn't needed any longer.